+++ src/render/site.js
@@ -96,7 +96,7 @@ export function renderSite(opts) {
 
   function addBlockTile(tabId, edit) {
     const tile = el('div', { class: 'ap-block', dataset: { span: '4' } }, [
-      el('button', { class: 'ap-addblock', onClick: () => edit.onAddBlock?.(tabId) }, [icon('plus'), 'Add Element']),
+      el('button', { class: 'ap-addblock', onClick: () => edit.onAddBlock?.(tabId) }, [icon('plus'), 'Add Section']),
     ]);
     return tile;
   }